+++ docs/about/changelog.md
@@ -20,6 +20,7 @@
       (`#msh preprocessor mode`)
     * Add command to set the shell to metaprogramming mode
       (`#msh metaprogram mode`)
+    * Add command to display a comment in the shell (`#msh echo`)
 
 * Changes to existing behaviour
     * File sections displayed by MDB now shows the source location of the